# api/openapi.yaml
openapi: 3.0.3
info:
  title: Club Management System API
  description: This is the API definition of the Club Management System.
  version: 0.0.0
servers:
  - url: /api/v1
    description: ""
paths:
...
  /letters/mail:
    post:
      operationId: sendMail
      tags:
        - letters
      summary: Send mail
      responses:
        "204":
          $ref: "#/components/responses/NoContent"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalServerErrror"
      security:
        - BearerJwt: []
      requestBody:
        required: true
        description: ""
        content:
          text/html:
            schema:
              type: object
      parameters: []
  /letters/pdf:
    post:
      operationId: getPdf
      tags:
        - letters
      summary: Get pdf
      responses:
        "200":
          description: The request was successful, and the server has returned the
            requested resource in the response body.
          content:
            application/pdf:
              schema:
                type: string
                format: binary
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalServerErrror"
      security:
        - BearerJwt: []
      parameters: []
      requestBody:
        required: true
        description: ""
        content:
          text/html:
            schema:
              type: object
...
components:
  responses:
    InternalServerErrror:
      description: The server encountered an unexpected condition that prevented it
        from fulfilling the request.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    BadRequest:
      description: The server could not understand the request due to invalid syntax.
        The client should modify the request and try again.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/BadRequestResponse"
...
    NoContent:
      description: The request was successful, but there is no content to return in
        the response.
    Unauthorized:
      description: Authentication is required to access the requested resource. The
        client must include the appropriate credentials.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: The server understood the request, but refuses to authorize it.
        Ensure the client has appropriate permissions.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  securitySchemes:
    BearerJwt:
      type: http
      scheme: Bearer
      bearerFormat: Json Web Token (JWT)
      description: "Jwt Auth: Authenticated requests contain a valid Json Web Token
        (JWT) as part of the `Authorization: Bearer <JWT>` header."
...
  schemas:
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
    BadRequestResponse:
      type: object
      properties:
        message:
          type: string
        errors:
          type: array
          items:
            $ref: "#/components/schemas/ErrorResponse"
      required:
        - message
...