# services/py-genai-helper/generated/models.py
# generated by datamodel-codegen:
#   filename:  openapi.yaml
#   timestamp: 2026-05-25T20:08:01+00:00

from __future__ import annotations
from pydantic import BaseModel, Field
from typing import Annotated


class ErrorResponse(BaseModel):
    message: str


class BadRequestResponse(BaseModel):
    message: str
    errors: list[ErrorResponse] | None = None


class Sport(BaseModel):
    name: str
    description: str
    created_at: str
    directors: list[str]


class SportCreate(BaseModel):
    name: str
    description: str | None = None
    directors: list[str] | None = None


class SportPartialUpdate(BaseModel):
    name: str | None = None
    description: str | None = None
    directors: list[str] | None = None


class Team(BaseModel):
    id: str
    name: str
    description: str
    created_at: str
    address: str
    sport: str
    trainers: list[str]
    trainees: list[str]


class TeamCreate(BaseModel):
    name: str
    description: str | None = None
    address: str | None = None
    sport: str
    trainers: list[str] | None = None
    trainees: list[str] | None = None


class TeamPartialUpdate(BaseModel):
    name: str | None = None
    description: str | None = None
    address: str | None = None
    sport: str | None = None
    trainers: list[str] | None = None
    trainees: list[str] | None = None


class Member(BaseModel):
    id: str
    first_name: str
    last_name: str
    email: str
    birthday: str
    phone_number: str
    address: str
    joining_date: str
    information: str


class MemberSummary(BaseModel):
    id: str
    first_name: str
    last_name: str
    email: str


class MemberPartialUpdate(BaseModel):
    first_name: str | None = None
    last_name: str | None = None
    email: str | None = None
    birthday: str | None = None
    phone_number: str | None = None
    address: str | None = None
    information: str | None = None


class MemberCreate(BaseModel):
    first_name: str
    last_name: str
    email: str | None = None
    birthday: str | None = None
    phone_number: str | None = None
    address: str | None = None
    information: str | None = None


class Event(BaseModel):
    id: str
    name: str
    description: str
    start_time: str
    end_time: str
    attendees: list[str] | None = None
    sports_linked: Annotated[
        list[str] | None,
        Field(description='Names of the sports associated with this event.'),
    ] = None
    teams_linked: Annotated[
        list[str] | None,
        Field(description='IDs of the teams associated with this event.'),
    ] = None
    creator: str


class EventSummary(BaseModel):
    id: str
    name: str
    start_time: str
    end_time: str


class EventPartialUpdate(BaseModel):
    name: str | None = None
    description: str | None = None
    start_time: str | None = None
    end_time: str | None = None
    attendees: list[str] | None = None
    sports_linked: list[str] | None = None
    teams_linked: list[str] | None = None


class EventCreate(BaseModel):
    name: str
    description: str | None = None
    start_time: str
    end_time: str
    attendees: list[str] | None = None
    sports_linked: list[str] | None = None
    teams_linked: list[str] | None = None


class Feedback(BaseModel):
    id: str
    event: str
    member: str
    creator: str
    created_at: str
    feedback: str


class FeedbackSummary(BaseModel):
    id: str
    event: str
    member: str
    creator: str
    created_at: str


class FeedbackPartialUpdate(BaseModel):
    event: str | None = None
    member: str | None = None
    feedback: str | None = None


class FeedbackCreate(BaseModel):
    event: str
    member: str
    feedback: str


class Transaction(BaseModel):
    id: str
    member: str
    creator: str
    amount_cents: int
    created_at: str
    title: str
    description: str


class TransactionPartialUpdate(BaseModel):
    member: str | None = None
    amount_cents: int | None = None
    title: str | None = None
    description: str | None = None


class TransactionCreate(BaseModel):
    member: str
    amount_cents: int
    title: str
    description: str | None = None


class Balance(BaseModel):
    member: str
    balance_cents: int