# services/spring-letter/src/generated/java/tum/devoops/letterservice/api/LettersApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.14.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package tum.devoops.letterservice.api;

import tum.devoops.letterservice.model.BadRequestResponse;
import tum.devoops.letterservice.model.ErrorResponse;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import io.swagger.v3.oas.annotations.media.ExampleObject;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.lang.Nullable;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", comments = "Generator version: 7.14.0")
@Validated
@Tag(name = "letters", description = "the letters API")
public interface LettersApi {

    default Optional<NativeWebRequest> getRequest() {
        return Optional.empty();
    }

    /**
     * POST /letters/pdf : Get pdf
     *
     * @param body  (required)
     * @return The request was successful, and the server has returned the requested resource in the response body. (status code 200)
     *         or The server could not understand the request due to invalid syntax. The client should modify the request and try again. (status code 400)
     *         or Authentication is required to access the requested resource. The client must include the appropriate credentials. (status code 401)
     *         or The server understood the request, but refuses to authorize it. Ensure the client has appropriate permissions. (status code 403)
     *         or The server encountered an unexpected condition that prevented it from fulfilling the request. (status code 500)
     */
    @Operation(
        operationId = "getPdf",
        summary = "Get pdf",
        tags = { "letters" },
        responses = {
            @ApiResponse(responseCode = "200", description = "The request was successful, and the server has returned the requested resource in the response body.", content = {
                @Content(mediaType = "application/pdf", schema = @Schema(implementation = org.springframework.core.io.Resource.class)),
                @Content(mediaType = "application/json", schema = @Schema(implementation = org.springframework.core.io.Resource.class))
            }),
            @ApiResponse(responseCode = "400", description = "The server could not understand the request due to invalid syntax. The client should modify the request and try again.", content = {
                @Content(mediaType = "application/pdf", schema = @Schema(implementation = BadRequestResponse.class)),
                @Content(mediaType = "application/json", schema = @Schema(implementation = BadRequestResponse.class))
            }),
            @ApiResponse(responseCode = "401", description = "Authentication is required to access the requested resource. The client must include the appropriate credentials.", content = {
                @Content(mediaType = "application/pdf", schema = @Schema(implementation = ErrorResponse.class)),
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            }),
            @ApiResponse(responseCode = "403", description = "The server understood the request, but refuses to authorize it. Ensure the client has appropriate permissions.", content = {
                @Content(mediaType = "application/pdf", schema = @Schema(implementation = ErrorResponse.class)),
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            }),
            @ApiResponse(responseCode = "500", description = "The server encountered an unexpected condition that prevented it from fulfilling the request.", content = {
                @Content(mediaType = "application/pdf", schema = @Schema(implementation = ErrorResponse.class)),
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            })
        },
        security = {
            @SecurityRequirement(name = "BearerJwt")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/letters/pdf",
        produces = { "application/pdf", "application/json" },
        consumes = { "text/html" }
    )
    
    default ResponseEntity<org.springframework.core.io.Resource> getPdf(
        @Parameter(name = "body", description = "", required = true) @Valid @RequestBody Object body
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"message\" : \"message\", \"errors\" : [ { \"message\" : \"message\" }, { \"message\" : \"message\" } ] }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"message\" : \"message\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"message\" : \"message\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"message\" : \"message\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * POST /letters/mail : Send mail
     *
     * @param body  (required)
     * @return The request was successful, but there is no content to return in the response. (status code 204)
     *         or The server could not understand the request due to invalid syntax. The client should modify the request and try again. (status code 400)
     *         or Authentication is required to access the requested resource. The client must include the appropriate credentials. (status code 401)
     *         or The server understood the request, but refuses to authorize it. Ensure the client has appropriate permissions. (status code 403)
     *         or The server encountered an unexpected condition that prevented it from fulfilling the request. (status code 500)
     */
    @Operation(
        operationId = "sendMail",
        summary = "Send mail",
        tags = { "letters" },
        responses = {
            @ApiResponse(responseCode = "204", description = "The request was successful, but there is no content to return in the response."),
            @ApiResponse(responseCode = "400", description = "The server could not understand the request due to invalid syntax. The client should modify the request and try again.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = BadRequestResponse.class))
            }),
            @ApiResponse(responseCode = "401", description = "Authentication is required to access the requested resource. The client must include the appropriate credentials.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            }),
            @ApiResponse(responseCode = "403", description = "The server understood the request, but refuses to authorize it. Ensure the client has appropriate permissions.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            }),
            @ApiResponse(responseCode = "500", description = "The server encountered an unexpected condition that prevented it from fulfilling the request.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            })
        },
        security = {
            @SecurityRequirement(name = "BearerJwt")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/letters/mail",
        produces = { "application/json" },
        consumes = { "text/html" }
    )
    
    default ResponseEntity<Void> sendMail(
        @Parameter(name = "body", description = "", required = true) @Valid @RequestBody Object body
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"message\" : \"message\", \"errors\" : [ { \"message\" : \"message\" }, { \"message\" : \"message\" } ] }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"message\" : \"message\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"message\" : \"message\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"message\" : \"message\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }

}